requirements.txt
streamlit==1.33.0
requests==2.31.0
python-dotenv==1.0.0
ollama==0.4.8
pandas==2.0.3
plotly==5.18.0
protobuf>=5.0,<6.0
opentelemetry-proto==1.32.1